Tool to share tags between mp3 files in directory
When ripping music from CDs, some times one or more of them will be
missing data that is contained on the rest of them.

Add a script to share this data (and fix the name) on files in the
same folder.

As this uses mutagen, it is licensed under the GPL v2